Add support for app layout initial height entry in YAML header
Controls with external CSS might render before the CSS loads, which
causes the height of the control to be incorrect in the app layout. We
should figure out how to delay rendering and size calculations until the
CSS loads (and also figure out why the external CSS in controls often
isn't getting cached), but for now we can manually specify the height of
the control so that the app layout gets the correct initial height. When
using this new height YAMl entry, check if the control's script is also
setting html.style.height (which is still relevant for document view);
if it is, change its value to match what you put in the YAML.