Update core-js version in package.json to match yarn.lock (for clarity) (#302)
My previous commit updated @babel/runtime-corejs3 version in
package.json, and I had meant to update the core-js version in
package.json along with it. This commit is just for clarity though,
since yarn.lock was already using the new version due to the fact that
yarn.lock was generated and added to the repo as part of my previous
commit.